refactor: sink URI codec primitives into engine-primitives to break execution_contract cycle
The encode/decode_uri_component[_strict] functions lived in
core::execution_contract, but core::artifact_ref imported them FROM there
while execution_contract imported ArtifactReference + schemes back FROM
artifact_ref -- a mutual cycle inside core.

Move the three pure-std URI codec fns down into
homeboy-engine-primitives::artifact_ref_scheme, right next to the
RUNNER_ARTIFACT/METADATA_ONLY scheme constants they operate on. Both
artifact_ref and execution_contract now depend on the primitives layer
instead of each other. execution_contract re-exports them so every
existing call site (including cross-crate agents/lab-runner) is unchanged.

This is the foundational cycle-break that unblocks a later full
execution-contract extraction (which additionally needs ArtifactReference
relocated off its observation dependency).
